+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.96.18] - 2026-03-26
+
+### Added
+
+- Add send button to ChatInput with new SVG icon and styles
+
+### Fixed
+
+- Change expandedMessages state from Map to Set
+- Update viewBox attribute in send.svg for proper alignment
+- Refactor padding calculations for map bounds into separate functions
+- Add primaryColor prop to ChatInput and pass it from ChatWindow
+
+### Changed
+
+- Rename handleChatSearchResults to handleChatLocations
+
 ## [1.96.17] - 2026-03-26
 
 ### Fixed